ecg_mise_en_prod/ia/requirements.txt
flask>=3.0,<4.0
numpy>=1.26,<2.2
tensorflow-cpu>=2.16,<2.21
